Added two-column layout
A two column layout system has been added to sub pages, simply add the
class two and the main container will sit to the left and get the
appropriate styles.

Then add a container-right id div into the wrapper above (or below) the
main container, this will then be the second column.