Contact Form - Beta 1
The contact form is out of alpha. Here is what it brings to the table
(All editions, JS, Non JS, PhP and server side. So the whole thing)

FOR JAVASCRIPT
- Instant verification of fields
- Smooth transitions
- Easy form submission
- Bounds detection ( keeps contact for on screen no matter what )
- Minimum width for mobile & desktop devices
- AJAX verification with SERVER side code.
- XSS verification
- Verify every field on submission (stop inspect element tampering)
- Easy to use interface
* TESTED ON:
* OS X Yosemite (10.10.3)
- Google Chrome ( 43.0.2357.124 ) *64-Bit
- Safari ( 8.0.5 )
* Windows 8.1 (64-Bit)
- Google Chrome (43.0.2357.124 m) *64-Bit
- Internet Explorer ( 11.0.2 )
- Firefox ( 38.0.5 )
- Firefox Developer Edition ( 39.0a2 )
A Browser passes a test if the website FUNCTIONS without being hindered.
Visual issues are omitted if they are not supported or do not work as
expected. For instance loss of blur when using Internet Explorer.
*NOT all browsers were up to date when testing.

FOR SERVER (PHP) & Javascript (CLIENT)
- Field verification
- Field XSS protected
- strip_tag before send
- Verification and content message sent (One to HEXCODE)
- 304 Returns
- Spam protection
- Remembers last message
- Blocks users on JAVASCRIPT from resending ALSO.
- Immediate verification

FOR SERVER (PHP) & Non Javscript users ( CLIENT )
- Same as above BUT;
Verification is NOT immediate (file reload)

Test contact form here:

* NO JS:
http://harryfelton.web44.net/digital_website/assets/server/contact_form.php
* JS: http://harryfelton.web44.net/digital_website/help.php
(CLICK CONTACT US)

*NOJS link will be changed on final release.